+++ src/benchmark_runners/non_learned_non_imv_sort_join_runner.cpp
@@ -127,8 +127,8 @@ void * non_learned_non_imv_sort_join_thread(void * param)
     
     auto partition_t1 = high_resolution_clock::now();
     auto partition_t2 = high_resolution_clock::now();
-    vector<uint32_t> curr_partition_timings_in_ms;
-    vector<uint32_t> final_partition_timings_in_ms;
+    vector<uint64_t> curr_partition_timings_in_ms;
+    vector<uint64_t> final_partition_timings_in_ms;
     for (int rp = 0; rp < RUN_NUMS; ++rp) 
     {
         //DEBUGMSG(1, "Thread-%d started running ... \n", my_tid);
@@ -155,7 +155,7 @@ void * non_learned_non_imv_sort_join_thread(void * param)
             deltaT = std::chrono::duration_cast<std::chrono::microseconds>(partition_t2 - partition_t1).count();
             //deltaT = (args->partition_end_time.tv_sec - args->start_time.tv_sec) * 1000000 + args->partition_end_time.tv_usec - args->start_time.tv_usec;
             printf("---- %5s partitioning costs time (ms) = %10.4lf\n", "Non-learned sort join", deltaT * 1.0 / 1000);
-            curr_partition_timings_in_ms.push_back((uint32_t)(deltaT * 1.0 / 1000)); //ms
+            curr_partition_timings_in_ms.push_back((uint64_t)(deltaT * 1.0 / 1000)); //ms
         }
     
         if(!(rp == (RUN_NUMS - 1))){
@@ -183,8 +183,8 @@ void * non_learned_non_imv_sort_join_thread(void * param)
     *************************************************************************/
     auto sorting_t1 = high_resolution_clock::now();
     auto sorting_t2 = high_resolution_clock::now();
-    vector<uint32_t> curr_sorting_timings_in_ms;
-    vector<uint32_t> final_sorting_timings_in_ms;
+    vector<uint64_t> curr_sorting_timings_in_ms;
+    vector<uint64_t> final_sorting_timings_in_ms;
     for (int rp = 0; rp < RUN_NUMS; ++rp) 
     {
         BARRIER_ARRIVE(args->barrier, rv);
@@ -206,7 +206,7 @@ void * non_learned_non_imv_sort_join_thread(void * param)
             deltaT = std::chrono::duration_cast<std::chrono::microseconds>(sorting_t2 - sorting_t1).count();
             //deltaT = (args->sort_end_time.tv_sec - args->partition_end_time.tv_sec) * 1000000 + args->sort_end_time.tv_usec - args->partition_end_time.tv_usec;
             printf("---- %5s sorting costs time (ms) = %10.4lf\n", "Non-learned sort join", deltaT * 1.0 / 1000);
-            curr_sorting_timings_in_ms.push_back((uint32_t)(deltaT * 1.0 / 1000)); //ms
+            curr_sorting_timings_in_ms.push_back((uint64_t)(deltaT * 1.0 / 1000)); //ms
         }
     
         if(!(rp == (RUN_NUMS - 1))){
@@ -271,8 +271,8 @@ void * non_learned_non_imv_sort_join_thread(void * param)
     Relation<KeyType, PayloadType> mergedRelS;
     auto merging_t1 = high_resolution_clock::now();
     auto merging_t2 = high_resolution_clock::now();
-    vector<uint32_t> curr_merging_timings_in_ms;
-    vector<uint32_t> final_merging_timings_in_ms;
+    vector<uint64_t> curr_merging_timings_in_ms;
+    vector<uint64_t> final_merging_timings_in_ms;
     for (int rp = 0; rp < RUN_NUMS; ++rp) 
     {
         BARRIER_ARRIVE(args->barrier, rv);
@@ -295,7 +295,7 @@ void * non_learned_non_imv_sort_join_thread(void * param)
             deltaT = std::chrono::duration_cast<std::chrono::microseconds>(merging_t2 - merging_t1).count();
             //deltaT = (args->multiwaymerge_end_time.tv_sec - args->sort_end_time.tv_sec) * 1000000 + args->multiwaymerge_end_time.tv_usec - args->sort_end_time.tv_usec;
             printf("---- %5s merging costs time (ms) = %10.4lf\n", "Non-learned sort join", deltaT * 1.0 / 1000);
-            curr_merging_timings_in_ms.push_back((uint32_t)(deltaT * 1.0 / 1000)); //ms
+            curr_merging_timings_in_ms.push_back((uint64_t)(deltaT * 1.0 / 1000)); //ms
         }
     
         if(!(rp == (RUN_NUMS - 1))){
@@ -334,8 +334,8 @@ void * non_learned_non_imv_sort_join_thread(void * param)
      *************************************************************************/
     auto join_t1 = high_resolution_clock::now();
     auto join_t2 = high_resolution_clock::now();
-    vector<uint32_t> curr_join_timings_in_ms;
-    vector<uint32_t> final_join_timings_in_ms;
+    vector<uint64_t> curr_join_timings_in_ms;
+    vector<uint64_t> final_join_timings_in_ms;
     for (int rp = 0; rp < RUN_NUMS; ++rp) 
     {
         BARRIER_ARRIVE(args->barrier, rv);
@@ -358,7 +358,7 @@ void * non_learned_non_imv_sort_join_thread(void * param)
             deltaT = std::chrono::duration_cast<std::chrono::microseconds>(join_t2 - join_t1).count();
             //deltaT = (args->mergejoin_end_time.tv_sec - args->multiwaymerge_end_time.tv_sec) * 1000000 + args->mergejoin_end_time.tv_usec - args->multiwaymerge_end_time.tv_usec;
             printf("---- %5s joining costs time (ms) = %10.4lf\n", "Non-learned sort join", deltaT * 1.0 / 1000);
-            curr_join_timings_in_ms.push_back((uint32_t)(deltaT * 1.0 / 1000)); //ms
+            curr_join_timings_in_ms.push_back((uint64_t)(deltaT * 1.0 / 1000)); //ms
         }
     
         if(!(rp == (RUN_NUMS - 1))){
@@ -379,7 +379,7 @@ void * non_learned_non_imv_sort_join_thread(void * param)
     }
 
     if(my_tid == 0){
-        std::vector<std::pair<std::string, std::vector<uint32_t>>> final_results =
+        std::vector<std::pair<std::string, std::vector<uint64_t>>> final_results =
          {{"Partition_in_ms", final_partition_timings_in_ms},
           {"Sort_in_ms", final_sorting_timings_in_ms},
           {"Merge_in_ms", final_merging_timings_in_ms},
