+++ src/benchmark_runners/non_partition_join_runner.cpp
@@ -1961,8 +1961,8 @@ void * npj_join_thread(void * param)
 #ifdef RUN_LEARNED_TECHNIQUES        
     auto learn_model_t1 = high_resolution_clock::now();
     auto learn_model_t2 = high_resolution_clock::now();
-    vector<uint32_t> curr_learn_model_timings_in_ms;
-    vector<uint32_t> final_learn_model_timings_in_ms;
+    vector<uint64_t> curr_learn_model_timings_in_ms;
+    vector<uint64_t> final_learn_model_timings_in_ms;
     for (int rp = 0; rp < RUN_NUMS; ++rp) 
     {
         if(args->tid == 0)
@@ -1985,7 +1985,7 @@ void * npj_join_thread(void * param)
             deltaT = std::chrono::duration_cast<std::chrono::microseconds>(learn_model_t2 - learn_model_t1).count();
             //deltaT = (t2.tv_sec - t1.tv_sec) * 1000000 + t2.tv_usec - t1.tv_usec;
             printf("---- Sampling and training models time (ms) = %10.4lf\n",  deltaT * 1.0 / 1000);
-            curr_learn_model_timings_in_ms.push_back((uint32_t)(deltaT * 1.0 / 1000)); //ms
+            curr_learn_model_timings_in_ms.push_back((uint64_t)(deltaT * 1.0 / 1000)); //ms
 
 #ifndef RUN_LEARNED_TECHNIQUES_WITH_FIRST_LEVEL_ONLY
             if(rp == RUN_NUMS - 1)
@@ -2050,10 +2050,10 @@ void * npj_join_thread(void * param)
     auto build_t2 = high_resolution_clock::now();
     ETHNonPartitionJoinBuild<KeyType, PayloadType> build_data; 
     
-    vector<uint32_t> final_build_timings_in_ms;
+    vector<uint64_t> final_build_timings_in_ms;
     for (int fid = 0; fid < npj_pf_num; ++fid) 
     {
-        vector<uint32_t> curr_build_timings_in_ms;
+        vector<uint64_t> curr_build_timings_in_ms;
         for (int rp = 0; rp < RUN_NUMS; ++rp) 
         {
             init_bucket_buffer(&overflowbuf);
@@ -2123,7 +2123,7 @@ void * npj_join_thread(void * param)
                 //deltaT = (args->partition_end_time.tv_sec - args->start_time.tv_sec) * 1000000 + args->partition_end_time.tv_usec - args->start_time.tv_usec;
                 deltaT = std::chrono::duration_cast<std::chrono::microseconds>(build_t2 - build_t1).count();
                 printf("---- %5s Build costs time (ms) = %10.4lf\n", npj_pfun[fid].fun_name, deltaT * 1.0 / 1000);
-                curr_build_timings_in_ms.push_back((uint32_t)(deltaT * 1.0 / 1000)); //ms
+                curr_build_timings_in_ms.push_back((uint64_t)(deltaT * 1.0 / 1000)); //ms
                 npj_total_num = 0;
                 npj_global_curse = 0;
             }
@@ -2150,10 +2150,10 @@ void * npj_join_thread(void * param)
     auto probe_t1 = high_resolution_clock::now();
     auto probe_t2 = high_resolution_clock::now();
 
-    vector<uint32_t> final_probe_timings_in_ms;
+    vector<uint64_t> final_probe_timings_in_ms;
     for (int fid = 0; fid < npj_pf_num; ++fid) 
     {
-        vector<uint32_t> curr_probe_timings_in_ms;
+        vector<uint64_t> curr_probe_timings_in_ms;
         for (int rp = 0; rp < RUN_NUMS; ++rp) 
         {
             BARRIER_ARRIVE(args->barrier, rv);
@@ -2177,7 +2177,7 @@ void * npj_join_thread(void * param)
                 deltaT = std::chrono::duration_cast<std::chrono::microseconds>(probe_t2 - probe_t1).count();
                 //deltaT = (args->end_time.tv_sec - args->partition_end_time.tv_sec) * 1000000 + args->end_time.tv_usec - args->partition_end_time.tv_usec;
                 printf("---- %5s Probe costs time (ms) = %10.4lf\n", npj_pfun1[fid].fun_name, deltaT * 1.0 / 1000);
-                curr_probe_timings_in_ms.push_back((uint32_t)(deltaT * 1.0 / 1000)); //ms
+                curr_probe_timings_in_ms.push_back((uint64_t)(deltaT * 1.0 / 1000)); //ms
             }
         }
 
@@ -2189,12 +2189,12 @@ void * npj_join_thread(void * param)
 
     if(args->tid == 0){
 #ifdef RUN_LEARNED_TECHNIQUES        
-        std::vector<std::pair<std::string, std::vector<uint32_t>>> final_results =
+        std::vector<std::pair<std::string, std::vector<uint64_t>>> final_results =
          {{"Learn_model_in_ms", final_learn_model_timings_in_ms},
           {"Build_in_ms", final_build_timings_in_ms},
           {"Probe_in_ms", final_probe_timings_in_ms}};
 #else
-        std::vector<std::pair<std::string, std::vector<uint32_t>>> final_results =
+        std::vector<std::pair<std::string, std::vector<uint64_t>>> final_results =
          {{"Build_in_ms", final_build_timings_in_ms},
           {"Probe_in_ms", final_probe_timings_in_ms}};
 #endif
